ml_models/cnn/requirements.txt
torch==2.2.2
torchvision==0.17.2
pandas==1.4.3
numpy==1.23.1
scikit-learn==1.1.1
mlflow==1.28.0
Pillow==9.2.0
matplotlib==3.5.2